+++ CHANGELOG.md
@@ -26,7 +26,7 @@ See for more information about XPrivacy 3 [this FAQ](https://github.com/M66B/XPr
 
 **Next release**
 
-* ...
+* Broadcast *biz.bokhorst.xprivacy.action.EXCEPTION* if the database could not be read ([issue](/../../issues/2081))
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)